Using size_t instead of int for array sizes and indexes (#10)

+++ src/main/c/functions.c
@@ -6,7 +6,7 @@
 /**
  * Returns the size of a triangular matrix, without the diagonal.
  */
-int size_triangular_matrix(const int n) {
+size_t size_triangular_matrix(const size_t n) {
   return ((n - 1) * n) / 2;
 }
 
@@ -27,8 +27,8 @@ double compute_overlap(const Particle p1, const Particle p2) {
 /**
  * Takes two arrays of vectors, and sums together the vectors with the same index.
  */
-void sum_vectors(const int size, const Vector *v1, const Vector *v2, Vector *result) {
-  for (int i = 0; i < size; ++i) {
+void sum_vectors(const size_t size, const Vector *v1, const Vector *v2, Vector *result) {
+  for (size_t i = 0; i < size; ++i) {
     result[i].x_component = v1[i].x_component + v2[i].x_component;
     result[i].y_component = v1[i].y_component + v2[i].y_component;
   }
@@ -38,8 +38,8 @@ void sum_vectors(const int size, const Vector *v1, const Vector *v2, Vector *res
  * Applies the forces to the particles with the same index,
  * and computes the resultant acceleration.
  */
-void compute_acceleration(const int size, const ParticleProperties *particles, const Vector *forces, Vector *resultant_accelerations){
-  for (int i = 0; i < size; ++i) {
+void compute_acceleration(const size_t size, const ParticleProperties *particles, const Vector *forces, Vector *resultant_accelerations){
+  for (size_t i = 0; i < size; ++i) {
     resultant_accelerations[i].x_component = forces[i].x_component / particles[i].mass;
     resultant_accelerations[i].y_component = forces[i].y_component / particles[i].mass;
   }
@@ -49,8 +49,8 @@ void compute_acceleration(const int size, const ParticleProperties *particles, c
  * Computes the displacement of the particles,
  * with an applied velocity for a given time delta.
  */
-void compute_displacement(const int size, const double dt, const Vector *velocities, Vector *displacements){
-  for (int i = 0; i < size; ++i) {
+void compute_displacement(const size_t size, const double dt, const Vector *velocities, Vector *displacements){
+  for (size_t i = 0; i < size; ++i) {
     displacements[i].x_component = displacements[i].x_component + velocities[i].x_component * dt;
     displacements[i].y_component = displacements[i].y_component + velocities[i].y_component * dt;
   }
@@ -59,8 +59,8 @@ void compute_displacement(const int size, const double dt, const Vector *velocit
 /** Derives the resultant velocity,
  * of an initial velocity with an applied acceleration for given a time delta.
  */
-void compute_velocity(const int size, const Vector *accelerations, const double dt, Vector *velocities){
-  for (int i = 0; i < size; ++i) {
+void compute_velocity(const size_t size, const Vector *accelerations, const double dt, Vector *velocities){
+  for (size_t i = 0; i < size; ++i) {
     velocities[i].x_component = velocities[i].x_component + accelerations[i].x_component * dt;
     velocities[i].y_component = velocities[i].y_component + accelerations[i].y_component * dt;
   }
@@ -72,13 +72,13 @@ void compute_velocity(const int size, const Vector *accelerations, const double 
  * Note: The size of the buffer should be equals to
  *       the size of a triangular matrix for the number of particles.
  */
-int compute_contacts(const int size, const Particle *particles, Contact *contacts_buffer) {
-  int k = 0;
+size_t compute_contacts(const size_t size, const Particle *particles, Contact *contacts_buffer) {
+  size_t k = 0;
   double overlap;
   Particle p1;
   Particle p2;
-  for (int i = 0; i < size; ++i) {
-    for (int j = i + 1; j < size; ++j) {
+  for (size_t i = 0; i < size; ++i) {
+    for (size_t j = i + 1; j < size; ++j) {
       p1 = particles[i];
       p2 = particles[j];
       overlap = compute_overlap(p1, p2);
